Consolidate the Bitflyer refresh HTTP setup
With feedback from Albert that he'd like to consolidate
the behavior into one file if possible, I've put the http call
method in a nested class and in the same file, for
 easier readability.
