+++ app/services/bitflyer/refresher.rb
@@ -2,8 +2,10 @@
 
 module Bitflyer
   class Refresher
+    REFRESH_TOKEN = "refresh_token"
+
     def self.build
-      new(impl_refresher: Bitflyer::Auth)
+      new(impl_refresher: RefresherHttpImpl)
     end
 
     def initialize(impl_refresher:)
@@ -21,13 +23,35 @@ module Bitflyer
 
       # Update with the latest Authorization
       bitflyer_connection.update!(
-        access_token: authorization.access_token,
-        refresh_token: authorization.refresh_token,
-        expires_in: authorization.expires_in,
-        access_expiration_time: authorization.expires_in.seconds.from_now
+        access_token: authorization['access_token'],
+        refresh_token: authorization['refresh_token'],
+        expires_in: authorization['expires_in'],
+        access_expiration_time: authorization['expires_in'].seconds.from_now
       )
       # Reload the model so consumers will have the most up to date information.
       bitflyer_connection.reload
+    end
+
+    class RefresherHttpImpl
+      # Public: Requests a refresh token from the Bitflyer /auth/token.
+      #
+      # token - The refresh token made from initial token authorization flow
+      #
+      # Returns an auth object
+      def self.refresh(token:, scope: Bitflyer.scope, http_client: Bitflyer::Auth.new)
+        # This is a temporary stop gap until this issue is addressed
+        # https://github.com/brave-intl/publishers/issues/2779
+
+        body = {
+          client_id: Bitflyer.client_id,
+          client_secret: Bitflyer.client_secret,
+          grant_type: REFRESH_TOKEN,
+          scope: scope,
+          refresh_token: token,
+        }
+        response = http_client.send(:post, Bitflyer.oauth_path, body)
+        JSON.parse(response.body)
+      end
     end
   end
 end
